refactor: finished working on reader architecture
