+++ text-analyzer/src/main/java/ru/itis/informatics/Main.java
@@ -1,20 +1,13 @@
 package ru.itis.informatics;
 
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.List;
-
-import ru.itis.informatics.reader.ReaderTXT;
+import ru.itis.informatics.reader.FileReader;
+import ru.itis.informatics.reader.extensions.ReaderTXT;
 
 public final class Main {
 	public static void main(String[] args) throws IOException {
-		Path path = Paths.get("resources/text.txt");
-		List<String> lines = Files.readAllLines(path);
-		
-		ReaderTXT readerTXT = new ReaderTXT();
+		final FileReader fileReader = new FileReader("resources/text.txt", new ReaderTXT());
 
-		System.out.println(readerTXT.getReadableLinesFrom(lines));
+		System.out.println(fileReader.getWords());
 	}
 }
