+++ text-analyzer/src/main/java/ru/itis/informatics/reader/extensions/ReaderTXT.java
@@ -1,4 +1,4 @@
-package ru.itis.informatics.reader;
+package ru.itis.informatics.reader.extensions;
 
 import java.util.ArrayList;
 import java.util.List;
